+++ src/main/java/de/l3s/learnweb/resource/speechRepository/SpeechRepositoryEntity.java
@@ -108,7 +108,7 @@ public class SpeechRepositoryEntity {
         int duration = 0;
         int multiply = 0;
         for (int i = tokens.length - 1; i >= 0; --i) {
-            duration += Integer.parseInt(tokens[i]) * Math.pow(60, multiply++);
+            duration += Integer.parseInt(tokens[i]) * (int) Math.pow(60, multiply++);
         }
 
         this.duration = duration;
